// app/v1/src/modules/location/enums/location.enums.ts
export enum WeekStartOptions {
    SUNDAY = "SUNDAY",
    MONDAY = "MONDAY",
    TUESDAY = "TUESDAY",
    WEDNESDAY = "WEDNESDAY",
    THURSDAY = "THURSDAY",
    FRIDAY = "FRIDAY",
    SATURDAY = "SATURDAY",
  }
  
  export enum LeaveQuotaResetOptions {
    ACCOUNTING_YEAR = "ACCOUNTING_YEAR",
    CALENDAR_YEAR = "CALENDAR_YEAR",
  }
  
  export enum Months {
    JANUARY = "JANUARY",
    FEBRUARY = "FEBRUARY",
    MARCH = "MARCH",
    APRIL = "APRIL",
    MAY = "MAY",
    JUNE = "JUNE",
    JULY = "JULY",
    AUGUST = "AUGUST",
    SEPTEMBER = "SEPTEMBER",
    OCTOBER = "OCTOBER",
    NOVEMBER = "NOVEMBER",
    DECEMBER = "DECEMBER",
  }
  